I've split the `CONV_2D` operation into two parts: im2col(`CONV_2D_STAGE_0`) and GEMM(`CONV_2D_STAGE_1`). I've enabled multi-threading for im2col and restructured the loops to make the cache more friendly. For clarity, I've used more intuitive and general variable names.

Taking the decode_first_stage phase in stable-diffusion as an example, when generating a 512x512 image.
For a detailed comparison, you can refer to the information provided here https://github.com/leejet/stable-diffusion.cpp/discussions/30

before:
```
perf_total_per_op_us[             ADD] = 1590.776 ms
perf_total_per_op_us[             MUL] = 708.338 ms
perf_total_per_op_us[          REPEAT] = 3385.188 ms
perf_total_per_op_us[      GROUP_NORM] = 754.513 ms
perf_total_per_op_us[         MUL_MAT] = 244.121 ms
perf_total_per_op_us[           SCALE] =  11.161 ms
perf_total_per_op_us[            CONT] =  16.623 ms
perf_total_per_op_us[         RESHAPE] =   0.111 ms
perf_total_per_op_us[         PERMUTE] =   0.003 ms
perf_total_per_op_us[        SOFT_MAX] =  12.856 ms
perf_total_per_op_us[         CONV_2D] = 43098.730 ms
perf_total_per_op_us[         UPSCALE] = 313.308 ms
perf_total_per_op_us[           UNARY] = 254.900 ms
```
after:
```
perf_total_per_op_us[             ADD] = 1595.238 ms
perf_total_per_op_us[             MUL] = 698.069 ms
perf_total_per_op_us[          REPEAT] = 3551.963 ms
perf_total_per_op_us[      GROUP_NORM] = 760.794 ms
perf_total_per_op_us[         MUL_MAT] = 219.111 ms
perf_total_per_op_us[           SCALE] =   9.338 ms
perf_total_per_op_us[            CONT] =  15.030 ms
perf_total_per_op_us[         RESHAPE] =   0.106 ms
perf_total_per_op_us[         PERMUTE] =   0.003 ms
perf_total_per_op_us[        SOFT_MAX] =  10.828 ms
perf_total_per_op_us[ CONV_2D_STAGE_0] = 5607.402 ms
perf_total_per_op_us[ CONV_2D_STAGE_1] = 12835.200 ms
perf_total_per_op_us[         UPSCALE] = 317.210 ms
perf_total_per_op_us[           UNARY] = 289.982 ms
```

The time for `CONV_2D` has been reduced from the original 43098.730 ms to 18442.602 ms (`CONV_2D_STAGE_0` + `CONV_2D_STAGE_1`).