@leejet 2023-08-27
By the way, I believe that the matrix multiplication used in both MUL_MAT and CONV_2D operations should be unified for easier optimization.