@ggerganov 2023-08-28
Nice speedup!

I am thinking how to avoid the 2-step process. Do you need it just because the INIT step is multi-threaded?
If it becomes multi-threaded, would you still need 2 stages?

I think it should be possible to reuse `ggml_compute_forward_mul_mat()` in stage 2, however, it does not support `src1` being F16. It has to be F32

Btw, I'm not 100% that the strategy of making the Conv2D into a matrix multiplication is the best thing to do. I initially did it like this since mul mat is familiar, but we waste a lot of extra memory this way. Maybe there is something better to do here - not sure. On the other hand, if we reuse the mul mat implementation it would be great because we won't have to write GPU code for Conv2D.

Will need some time to figure out how to refactor and merge this.
@slaren Let me know if you get any ideas how to improve this