@leejet 2023-08-28
> I am thinking how to avoid the 2-step process. Do you need it just because the INIT step is not multi-threaded?
If it becomes multi-threaded, would you still need 2 stages?

If the INIT step is multithreaded, then indeed there might be no need to split it into two stages.

> Btw, I'm not 100% that the strategy of making the Conv2D into a matrix multiplication is the best thing to do.

Im2col is a common practice in many frameworks like Caffe, PyTorch, and others. However, these frameworks often implement other optimization techniques such as Winograd, where they choose different calculation methods based on the input format.