@ggerganov 2023-09-08
I think for now we'll merge this as proposed.

At some point, we have to make Stage 1 reuse the matrix multiplication operator because this will bring massive speed-up on Apple Silicon via the Accelerate framework.

@leejet Let me know if this is still the version you want to have upstream.