@AngryLoki 2023-09-13
@leejet , could you check, please, if there is a race condition? I tried to compare results of old conv2d with new gemm-based and got same values with single-threaded, but different with multithreaded (and I know for sure that exactly old code was correct).

```
    int64_t ne00 = 3, ne01 = 3, ne02 = 640, ne03 = 640;
    int64_t ne10 = 80, ne11 = 80, ne12 = 640, ne13 = 1;

    int s0 = 1;
    int s1 = 1;
    int p0 = 1;
    int p1 = 1;
    int d0 = 1;
    int d1 = 1;

    std::vector<float> adata(ne00 * ne01 * ne02 * ne03);
    for (size_t i = 0; i < adata.size(); i++) adata[i] = 1;

    std::vector<uint16_t> hadata(ne00 * ne01 * ne02 * ne03);
    ggml_fp32_to_fp16_row(adata.data(), hadata.data(), adata.size());

    std::vector<float> bdata(ne10 * ne11 * ne12 * ne13);
    for (size_t i = 0; i < bdata.size(); i++) bdata[i] = 1;

    struct ggml_init_params params_ctx = {
        .mem_size = 200 * 1024 * 1024,
        .mem_buffer = NULL,
        .no_alloc = false,
    };

    struct ggml_context* ctx = ggml_init(params_ctx);

    struct ggml_tensor* a = ggml_new_tensor_4d(ctx, GGML_TYPE_F32, ne00, ne01, ne02, ne03);
    memcpy(a->data, adata.data(), adata.size() * sizeof(adata[0]));
    struct ggml_tensor* ha = ggml_new_tensor_4d(ctx, GGML_TYPE_F16, ne00, ne01, ne02, ne03);
    memcpy(ha->data, hadata.data(), hadata.size() * sizeof(hadata[0]));

    struct ggml_tensor* b = ggml_new_tensor_4d(ctx, GGML_TYPE_F32, ne10, ne11, ne12, ne13);
    memcpy(b->data, bdata.data(), bdata.size() * sizeof(bdata[0]));

    struct ggml_tensor* r1 = ggml_conv_2d_stage_0(ctx, ha, b, s0, s1, p0, p1, d0, d1); // [N, OH, OW, IC * KH * KW]
    struct ggml_tensor* ref1 = ggml_conv_2d_stage_1(ctx, ha, r1);

        struct ggml_compute_params params = {
            .type = GGML_TASK_INIT,
            .ith = 0,
            .nth = 1,
            .wsize = 0,
            .wdata = NULL,
        };

        ggml_compute_forward_conv_2d_stage_0(&params, ha, b, r1);

        params.type = GGML_TASK_COMPUTE;
        #pragma omp parallel
        {
            params.nth = omp_get_num_threads();
            params.ith = omp_get_thread_num();
            ggml_compute_forward_conv_2d_stage_0(&params, ha, b, r1);
        }

        #pragma omp parallel
        {
            params.nth = omp_get_num_threads();
            params.ith = omp_get_thread_num();
            ggml_compute_forward_conv_2d_stage_1(&params, ha, r1, ref1);
        }

    struct ggml_tensor* ref2 = ggml_new_tensor_4d(ctx, GGML_TYPE_F32, ref1->ne[0], ref1->ne[1], ref1->ne[2], ref1->ne[3]);
    ref2->op_params[0] = ref2->op_params[1] = ref2->op_params[2] = ref2->op_params[3] = ref2->op_params[4] = ref2->op_params[5] = 1;
    void *buf = malloc(200*1024*1024);

        struct ggml_compute_params params = {
            .type = GGML_TASK_INIT,
            .ith = 0,
            .nth = 1,
            .wsize = 200*1024*1024,
            .wdata = buf,
        };

        ggml_compute_forward_conv_2d_f16_f32(&params, ha, b, ref2);

        params.type = GGML_TASK_COMPUTE;
        #pragma omp parallel
        {
            params.nth = omp_get_num_threads();
            params.ith = omp_get_thread_num();
            ggml_compute_forward_conv_2d_f16_f32(&params, ha, b, ref2);
        }

    const float *dref1 = (float *)(ref1->data);
    printf("new: %f %f %f %f %f %f %f\n", dref1[0], dref1[1], dref1[2], dref1[3], dref1[4], dref1[5], dref1[6]);

    const float *dref2 = (float *)(ref2->data);
    printf("old: %f %f %f %f %f %f %f\n", dref2[0], dref2[1], dref2[2], dref2[3], dref2[4], dref2[5], dref2[6]);
```

```
new: 2240.000000 3360.000000 3360.000000 3360.000000 3360.000000 3360.000000 3360.000000
old: 2560.000000 3840.000000 3840.000000 3840.000000 3840.000000 3840.000000 3840.000000
```

where new is randomly 2560,... or 2240,..., or even 0.000000,...  only when omp is enabled. In single-thread results are the same.